hw/char/fsl-linflex: don't update irq on read
Attempting (again) to fix the issue with serial device
misbehaving after machine has been booted for some time.

This change removes the irq update on register read.
This seems to have resolved the issue with the missing
(or extra?) interrupt.

Copy pasting into the uart termianl still triggers
this behavior.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>