nxp-s32g: Add another clock for can controller
This change renames the system default clock to ipg and
creates another clock, per, to use as input for the CAN
controller.

This change is needed by vxworks, since vxworks flexcan
driver specifically looks for "per" clock node in the
device tree, instead of looking for the clock phandle
specified in the flexcan dt node.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>