hw/net/fsl_flexcan: make MB array 32 bits wide
The backing array for the message buffer mmio region was
originally allocated as a uint8 array instead of uint32
because the message buffer is implemented as memory mapped
embedded RAM, and not registers. However, accesses and
fields within the Message Buffer space is aligned to 32
bits. To make parsing the fields from the MB control
registers easier down the line, change the backing array
to uint32_t so we access MB at 4 byte alignment.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>