contrib/plugins: new plugin to track device access
The devaccess tcg plugin is intended to track and record
details of each memory operation that translates into an
io operation on a memory mapped device. There is some
overlap in functionality with the hwprofile plugin, but
this is intended to target a more specific use case, and
allow us the opportunity to optimize the trace output
logic in the future.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>