hw/arm/wr-arm: make cpu type configurable
Currently the wr-arm machine is hardcoded to use cortex-a53
as its cpu type. qemu implements a -cpu option when invoking
qemu, allowing the user to specify the specific cpu type to
emulate.

This change allows the user to specify their arm cpu of choice
to better match the platform they are trying to emulate. If
not provided by the user, the machine defaults to cortex-a53.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>