mmu-500: init iommu memory regions
The upstream does not have the arm_generic_fdt machine
and supporting bits. Remove the parse_reg code that
exists as part of the fdt-generic-util infrastructure
that makes possible the use of this device with the
-hw-dtb option on Xilinx fork.

Initialize the iommu memory region in SMMU.tbu during
smmu_init. This was originally done as part of
smmu_parse_reg, but since this is no longer invoked as
part of fdt_init_qdev, we must do it elsewhere. Also,
remove the MemoryRegion member of TBU struct (TBU.mr),
since this appears to be redundant next to the iommu
memory region located at TBU.iommu. Rather than use
TBU.mr as the memory region root for TBU.as address
space, use TBU.iommu as the memory region. I admit I
do not fully understand how TBU.mr was initialized
in the arm-generic-fdt machine, so it is possible
removal of TBU.mr was done in error.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>