mmu-500: split device state into header file
The xilinx implementation of smmu500 was intended to be used
with the fdt-driven generic arm machine. As such, the mmu-500
device model was not realized as a static member of any
specific machine implementation. Rather, it was created as an
instance of FDTGenericMMapClass and initialized through the
parse_reg recursion terminating at sysbus_parse_reg and
dynamically attached to the target machine in fdt_init_qdev.

This change creates a new header file for the mmu-500
containing the SMMU state structure as well as the simple qom
object type structure.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>