versal-virt: add pcie device tree nodes
Create device tree nodes for the GPEX PCIe host bridge
configuration space "pci-host-ecam-generic" for the
xlnx-versal-virt machine.

When VxWorks parses the pcie interrupt-map property, it looks to
the "address-cells" property of the interrupt parent to determine
how many cells are in each row of the interrupt-map, which affects
how it parses the SPI/PPI and trigger type of the legacy intx
interrupts. The address-cells properties are not inherited from
ancestors in the device tree, so must be explicitly defined in the
gic node. So we also make this change here.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>