versal: add size-cells prop to gic fdt node
size-cells property is not inherited from ancestors
in the device tree, so must be explicitly defined in
the GIC node.

Devices that specify the GIC as the interrupt parent
format the size field in their reg property according
to the value of the GIC size-cells property.

Without the size-cells property in the GIC node, the
reg property of the pcie controller and other devices
using the GIC as the interrupt parent may be parsed
incorrectly, resulting in mmio regions being mapped
incorrectly.

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>