versal-virt: move GIC node to top of device tree
VxWorks does only one pass over the device tree and treats
the GIC the same as any other peripheral in terms of
initialization priority. As a result, any devices that use
dynamic interrupts that specify the GIC as the interrupt
parent need to be parsed and initialized after the GIC.

To accomodate the VxWorks behavior, place the GIC node in
the device tree such that it precedes any other peripherals.

Note: Devices that use static interrupts (PPI/SPI) do not
have the same requirement, as VxWorks has a mechanism to
defer connection of statically assigned interrupts to the
interrupt initialization finalize routine.
See vxbIntDeferFinalize()

Signed-off-by: Nelson Ho <Nelson.Ho@windriver.com>