imx8mp-evk: fdt add ccm clock specifiers
Change ccm clock-cells property from 0 to 1.
This means that any devices that specify the ccm as
clock input need to provide a (clk_phandle, clk_specifier)
as the value to the clocks property.

This change is kind of a mess, since we don't have a good
understanding of what the different clock specifiers mean
to the hardware. But the VxWorks ccm and enet drivers are
not happy unless we provide the clock specifier for all
clock inputs.

Signed-off-by: Nelson Ho <nelson.ho@windriver.com>