callstack plugin: avoid reading link register
Currently we read the link register (x30) contents when we
detect a RET instruction, so that we can match the return
to the correct stack frame to pop. Register reading is
extremely slow.

To avoid having to read the link register contents, we can
instead read the program counter of the instruction that
immediately follows the RET instruction. To achieve this
though, the vcpu_insn_exec callback must be run after
EVERY instruction. However, this is still faster than
having to read register x30 at every single RET instruction.